Added endhints plugin
This plugin changes the position of the grey hints to
change from in the middle of the text to the edge of
the line, helping in getting a cleaner look for reading
the code. Hints in the middle of the line break the
indentation flow and make's it confusing while writing code.